loading and saving the world is now working perfectly

+++ Meat.cpp
@@ -1,10 +1,10 @@
 #include "Meat.h"
 
-Meat::Meat(int x, int y, int radius, int quantity) : Resource(x, y, radius, quantity, EVOLUTION_MEAT)
+Meat::Meat(double x, double y, double radius, int quantity) : Resource(x, y, radius, quantity, EVOLUTION_MEAT)
 {
 }
 
-Meat::Meat(Coordinate c, int radius, int quantity) : Resource(c,radius,quantity, EVOLUTION_MEAT)
+Meat::Meat(Coordinate c, double radius, int quantity) : Resource(c,radius,quantity, EVOLUTION_MEAT)
 {
 }
