+++ Meat.h
@@ -6,8 +6,8 @@
 class Meat : public Resource
 {
 public:
-  Meat(int x, int y, int radius, int quantity);
-  Meat(Coordinate c, int radius, int quantity);
+  Meat(double x, double y, double radius, int quantity);
+  Meat(Coordinate c, double radius, int quantity);
   virtual unsigned int getTypeId() const { return ID_MEAT; }
   int getNeralNetworkId() const { return NN_ID_MEAT; }
   int eat(int quantity);
